// packages/sumsub/src/types.ts
export type ApplicantReviewStatus =
  | "init"
  | "pending"
  | "prechecked"
  | "queued"
  | "completed"
  | "onHold";

export type ApplicantReviewAnswer =
  | "GREEN"
  | "RED";

export interface ApplicantReviewResult {
  reviewAnswer?: ApplicantReviewAnswer;
  rejectLabels?: string[];
  reviewRejectType?: "FINAL" | "RETRY";
  moderationComment?: string;
  clientComment?: string;
}
// ...
export type SumsubWebhookType =
  | "applicantReviewed"
  | "applicantPending"
  | "applicantCreated"
  | "applicantOnHold"
  | "applicantActionPending"
  | "applicantActionReviewed"
  | "applicantActionOnHold"
  | "applicantReset"
  | "applicantPersonalInfoChanged"
  | "applicantPrechecked"
  | "videoIdentStatusChanged"
  | "applicantDeleted"
  | "applicantLevelChanged"
  | "applicantWorkflowCompleted"
  | string;

export interface SumsubWebhookPayload {
  applicantId: string;
  inspectionId?: string;
  correlationId?: string;
  externalUserId?: string;
  type: SumsubWebhookType;
  reviewResult?: ApplicantReviewResult;
  reviewStatus?: ApplicantReviewStatus;
  createdAtMs?: string;
  sandboxMode?: boolean;
  [key: string]: unknown;
}